refactor conftest as tests helper
