+++ astropy/cosmology/tests/conftest.py
@@ -1,0 +1,6 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+
+"""Configure the tests for :mod:`astropy.cosmology`."""
+
+from astropy.cosmology.tests.helper import clean_registry
+from astropy.tests.helper import pickle_protocol
