+++ astropy/cosmology/tests/helper.py
@@ -1,6 +1,9 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-"""Configure the tests for :mod:`astropy.cosmology`."""
+"""
+This module provides the tools used to internally run the cosmology test suite
+from the installed astropy.  It makes use of the `pytest`_ testing framework.
+"""
 
 ##############################################################################
 # IMPORTS
@@ -13,23 +16,32 @@ import pytest
 
 # LOCAL
 from astropy.cosmology import core
-from astropy.tests.helper import pickle_protocol  # noqa: F401
+
+__all__ = ["get_redshift_methods", "clean_registry"]
 
 ###############################################################################
 # FUNCTIONS
 
 
-def get_redshift_methods(cosmology, allow_private=True, allow_z2=True):
+def get_redshift_methods(cosmology, include_private=True, include_z2=True):
     """Get redshift methods from a cosmology.
 
     Parameters
     ----------
     cosmology : |Cosmology| class or instance
+    include_private : bool
+        Whether to include private methods, i.e. starts with an underscore.
+    include_z2 : bool
+        Whether to include methods that are functions of 2 (or more) redshifts,
+        not the more common 1 redshift argument.
 
     Returns
     -------
     set[str]
+        The names of the redshift methods on `cosmology`, satisfying
+        `include_private` and `include_z2`.
     """
+    # Get all the method names, optionally sieving out private methods
     methods = set()
     for n in dir(cosmology):
         try:  # get method, some will error on ABCs
@@ -38,7 +50,7 @@ def get_redshift_methods(cosmology, allow_private=True, allow_z2=True):
             continue
 
         # Add anything callable, optionally excluding private methods.
-        if callable(m) and (not n.startswith('_') or allow_private):
+        if callable(m) and (not n.startswith('_') or include_private):
             methods.add(n)
 
     # Sieve out incompatible methods.
@@ -60,7 +72,7 @@ def get_redshift_methods(cosmology, allow_private=True, allow_z2=True):
         elif len(params) >= iz1 + 1 and not params[iz1].startswith("z"):  # First non-self arg is z.
             methods.discard(n)
         # If methods with 2 z args are not allowed, the following arg is checked.
-        elif not allow_z2 and (len(params) >= iz1 + 2) and params[iz1 + 1].startswith("z"):
+        elif not include_z2 and (len(params) >= iz1 + 2) and params[iz1 + 1].startswith("z"):
             methods.discard(n)
 
     return methods
@@ -71,6 +83,7 @@ def get_redshift_methods(cosmology, allow_private=True, allow_z2=True):
 
 @pytest.fixture
 def clean_registry():
+    """`pytest.fixture` for clearing and restoring ``_COSMOLOGY_CLASSES``."""
     # TODO! with monkeypatch instead for thread safety.
     ORIGINAL_COSMOLOGY_CLASSES = core._COSMOLOGY_CLASSES
     core._COSMOLOGY_CLASSES = {}  # set as empty dict
