+++ docs/changes/cosmology/12966.feature.rst
@@ -1,0 +1,2 @@
+The new module ``cosmology/tests/helper.py`` has been added to provide tools
+for testing the cosmology module and related extensions.
